Make marker sizing more consistent; given in physical points (1/72 inch)
In the new scheme, the marker size is specified in as the nominal diameter of the mark in points (1/72 of a inch). If the GKS window has not been resized, they will be rendered accurately at this physical size. If the window is resized, they will be rescaled by a zoom factor, along with line widths and everything else. Marker type 1 is still rendered at a size of one screen unit (Apple 'points', not physical points) no matter what. I might change this to include the zoom factor at a later time.