Make window resize immediate, not stored in display list
This solves the problem with corrupted displays when DL contains
SET_WS_VIEWPORT. It doesn't make sense for this to be in the DL as
it needs to trigger a redraw anyway.

Remaining issue: no way to postpone redraw after resize when subsequent
update will make it redundant. Will result in flickering.