Acheive correctly sized rendering in at least one SVG agent
Ok, this looks a bit weird. Let me explain. SVG dimensions given without
units are assumed to be in 'CSS 'pixels'. CSS pixels are not device
pixels. They are a physical unit of length, 96 to the inch. It should be
sufficient to give the width and height in any unit: CSS pixels, inches,
cm etc. Through experimentation, I found that giving the size with units
(inches or points) AND the viewBox in pixels was sufficient to get a
correctly sized PDF rendering using the library/Python package cairosvg.